refactor sync rag systems
