+++ src/fed_rag/_bridges/llamaindex/_managed_index.py
@@ -20,7 +20,7 @@ from llama_index.core.schema import Node as LlamaNode
 from llama_index.core.schema import NodeWithScore, QueryBundle
 
 from fed_rag.base.generator import BaseGenerator
-from fed_rag.core._rag_system import SourceNode, _RAGSystem
+from fed_rag.core.rag_system._sync import SourceNode, _RAGSystem
 from fed_rag.data_structures import KnowledgeNode
 from fed_rag.exceptions import BridgeError
