+++ src/fed_rag/core/no_encode_rag_system/__init__.py
@@ -1,0 +1,3 @@
+from .sync import NoEncodeRAGSystem
+
+__all__ = ["NoEncodeRAGSystem"]
