+++ src/fed_rag/core/no_encode_rag_system/_sync.py
