+++ src/fed_rag/core/rag_system/__init__.py
@@ -1,0 +1,3 @@
+from .sync import RAGSystem
+
+__all__ = ["RAGSystem"]
