+++ src/fed_rag/core/rag_system/_sync.py
