+++ tests/bridges/test_llamaindex.py
@@ -12,7 +12,7 @@ from fed_rag._bridges.llamaindex._managed_index import (
     convert_source_node_to_llama_index_node_with_score,
 )
 from fed_rag._bridges.llamaindex.bridge import LlamaIndexBridgeMixin
-from fed_rag.core._rag_system import KnowledgeNode, SourceNode, _RAGSystem
+from fed_rag.core.rag_system._sync import KnowledgeNode, SourceNode, _RAGSystem
 from fed_rag.exceptions import BridgeError
